Refactor watson so use promises
