+++ app/utils/socketio.js
@@ -3,13 +3,14 @@ const tmi = require('tmi.js');
 const chooseChannel = require('./tmiConfig');
 
 const eventEmitter = require('./eventEmitter');
-const analyzer = require('../watson/analyzer');
+const storedChannel = require('./storedChannels');
 
 
 // Listenes for a new chat message and moves the message into the analyzer file
-eventEmitter.on('chatMessage', (msgObj) => {
-  analyzer.intoTones(msgObj);
-});
+// eventEmitter.on('chatMessage', (msgObj) => {
+//   channel = msgObj.channel;
+//   storedChannel.intoTones(msgObj);
+// });
 
 exports.establishConnection = () => {
   let twitchClient = null;
@@ -63,10 +64,10 @@ exports.ioConnect = server => {
       io.to(channel).emit('message', message);
     });
 
-    eventEmitter.on('toneUpdate', tone => {
-      let channel = tone.channel.substr(1);
-      io.to(channel).emit('tone:update', tone);
-    });
+    // eventEmitter.on('toneUpdate', tone => {
+    // let channel = tone.channel.substr(1);
+    //   io.to(channel).emit('tone:update', tone);
+    // });
 
   });
