+++ app/watson/analyzer.js
@@ -1,8 +1,7 @@
 const watson = require('watson-developer-cloud');
 const toneKey = require('../.appkeys.json');
-const toneController = require('../db/controllers/toneController');
 
-const tone_analyzer = watson.tone_analyzer({
+const toneAnalyzer = watson.tone_analyzer({
   username: toneKey.tone_analyzer.username,
   password: toneKey.tone_analyzer.password,
   version: 'v3-beta',
@@ -10,35 +9,19 @@ const tone_analyzer = watson.tone_analyzer({
   sentences: 'false',
 });
 
-const storedMessages = [];
-let channel = '';
-
-// Takes in the chat message, puts it into the storedMessages array
-exports.intoTones = (message) => {
-  channel = message.channel;
-  storedMessages.push(message.msg);
-};
-
-// Timer for pushing messages to watson for analysis
-setInterval(() => {
-  if (storedMessages.length) {
-    const arrayToText = { text: "'" + storedMessages.join('') + "'" };
-    getTone(arrayToText);
-    storedMessages.splice(0, storedMessages.length);
-  }
-}, 3000);
-
-// Passes string text to watson for analysis
-function getTone(message) {
-  tone_analyzer.tone(message, (err, tone) => {
-    if (err) {
-      console.log(err);
-    } else {
-      // console.log(JSON.stringify(tone, null, 2));
-      toneCallback(tone);
-    }
+// Returns a promise for the watson analysis
+exports.runAnalysis = (messages) => {
+  return new Promise((resolve, reject) => {
+    // Passes string text to watson for analysis
+    toneAnalyzer.tone(messages, (err, tone) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(toneCallback(tone));
+      }
+    });
   });
-}
+};
 
  // Converts a tone category into a flat object with tone values
 function getToneValues(toneCategory) {
@@ -65,7 +48,6 @@ function getTones(tone) {
 function toneCallback(data) {
   const tone = {
     toneData: {},
-    channel,
   };
 
   // Results for the updated full transcript's tone
@@ -76,6 +58,6 @@ function toneCallback(data) {
   //     tone.sentence = getTones(data.sentences_tone[data.sentences_tone.length - 1]);
 
   // console.log('inside toneCallback', tone);
-  // calls rethinkdb and saves the tone results
-  toneController.saveTone(tone);
+  // returns to the promise with a flattened object
+  return tone;
 }
